+++ internal/models/like.go
@@ -5,6 +5,6 @@ import "github.com/google/uuid"
 // easyjson -all ./internal/models/like.go
 
 type Like struct {
-	LikesCount int       `json:"likes_count"`
+	LikesCount int64     `json:"likes_count"`
 	PostID     uuid.UUID `json:"post_id"`
 }
